Version 1.1
Now when the cost of services for a car is calculated, current appointments for that car will be deleted;
Changed the type of all primary keys from int to long.

+++ src/main/java/classes/car/Car.java
@@ -15,7 +15,7 @@ public class Car
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Id
     @Column(name = "id")
-    private int m_id;
+    private long m_id;
 
     @Basic
     @Column(name = "plate_number")
@@ -55,9 +55,9 @@ public class Car
         m_fuel = fuel;
     }
 
-    public int getId() { return m_id; }
+    public long getId() { return m_id; }
 
-    public void setId(int m_id) { this.m_id = m_id; }
+    public void setId(long m_id) { this.m_id = m_id; }
 
     public String getPlateNumber() { return m_plate_number; }
